Added test cases for validity checks on validated_cols.
Moved old tests into new file because they only tested the
_unify_timestamps method.

Refs #26
